If you spend too much time thinking about a thing, you'll never get it done

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -2,21 +2,22 @@
 #include "function_pointers.h"
 
 /**
- * array_iterator - Executes a function on each element of an array
+ * array_iterator - Executing  a function on each array element
+ *
  * @array: The array to iterate over
  * @size: The size of the array
  * @action: A pointer to the function to execute
  */
+
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
 	int i;
 	if (array == NULL || action == NULL)
-	return;
+		    return;
 
-		for (i = 0; i < size; i++)
-	
+	for (i = 0; i < (int)size; i++) 
 	{
-		action(array[i]);
+	action(array[i]);
 	}
 }
